(navbar-item): move navigation main to main folder and solve story book
