+++ src/components/navigation/main/navigation-item/NavigationItem.scss
