+++ src/stories/navigation/navigation-item/NavigationItem.children.stories.ts
@@ -1,7 +1,7 @@
 import { setup, type Meta, type StoryObj } from "@storybook/vue3";
 import { action } from '@storybook/addon-actions';
 import ClickOutside from "@ui/utilities/directives/clickOutside";
-import NavigationItem from '@/components/navigation/navigation-item/NavigationItem.vue';
+import NavigationItem from '@/components/navigation/main/navigation-item/NavigationItem.vue';
 import BaseIcon from "@/components/base/base-icon/BaseIcon.vue";
 
 setup((app) => app.directive('click-outside', ClickOutside));
